requirements.txt
Flask==3.0.3
gunicorn==22.0.0
uvicorn[standard]==0.30.5
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
python-dotenv==1.0.1
httpx==0.27.2
pandas==2.2.2
pyarrow==17.0.0

numpy==1.26.4
scikit-learn==1.5.1
statsmodels==0.14.2
